AI DASHBOARD/requirements.txt
streamlit==1.31.0
google-generativeai==0.3.2
pandas==2.1.4
plotly==5.18.0
python-dotenv==1.0.0
Pillow==10.2.0


